chore: migrate expressionrouter tests to kongintegration
